requirements.txt
nltk
contractions
pyspellchecker
pandas
scikit-learn
python-docx